move sampling operations to dedicated file + begin sampling expansion
